Add description of non-0 to non-0 allowance change to Lorentz FA1.2
Problem: The description of Approve in Lorentz's FA1.2 does not mention
the failure that will happen when trying to directly change an allowance
from a non-zero value to another non-zero value, nor the reason for it.

Solution: add an example of said failure and a brief explanation of
the reason for it as well as what to do to perform such change.