Update macOS installation instructions
Problem: Currently tezos-packaging repository has two mirrors to provide
formulae for the latest stable and RC versions of the Octez.
The main repository doesn't provide any guarantees on which version will
be installed from it. Such behaviour can be confusing for users.

Solution: Use stable mirror repository in the instructions and mention
that one can install RC as well.